+++ src/main/java/com/example/daobe/auth/application/TokenExtractor.java
@@ -1,4 +1,4 @@
-package com.example.daobe.auth.service;
+package com.example.daobe.auth.application;
 
 public interface TokenExtractor {
